Janky workaround for toggle...
So previous edit worked but disabled the functionality of the button. Now clicking anywhere in the 'page-content-wrapper' div will open / close the sidebar. I removed the button click function but moved the button into that wrapper so it considers it part of it and functions as a button. 

Idk. Will look into this later. Essentially need a way to check if sidebar is open THEN use this click page-content-wrapper to close otherwise use button.