Updated UI, fixed bugs
To allow for a cleaner UI a dismissible well has been added that alerts the user to how to access the weather options. 
This well is inside the content div and so on close it will automatically open the weather options for viewing.

Fixed an issue where the icon would not disappear during reloading of weather 
Increased icon size to 100px on display

Improved alert for failure to retrieve location data (with steps for user to rectify) 
Improved alert for failure to retrieve Flickr image

Looked into issue of Flickr finding an image but having it replaced with a stock 'no longer available' image. No solution yet...

Future features:
Display artist, title, and link to image for Flickr and stock background photos
Allow cycling through multiple Flickr images (when available)
Allow user input of location (complex will have to look further into the approach to this)
Bring in improved icons because the current ones pulled from Apixu are small and have poor aesthetic